Refactor to avoid violating sealed packages.
