+++ .idea/runConfigurations/Run_Integration_Tests.xml
@@ -1,28 +1,20 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Run Integration Tests" type="JUnit" factoryName="JUnit">
-    <extension name="coverage" enabled="false" merge="false" sample_coverage="true" runner="idea">
+    <module name="cordapp_integrationTest" />
+    <extension name="coverage">
       <pattern>
-        <option name="PATTERN" value="com.template.*" />
+        <option name="PATTERN" value="com.template.contracts.*" />
         <option name="ENABLED" value="true" />
       </pattern>
     </extension>
-    <module name="cordapp_integrationTest" />
-    <option name="ALTERNATIVE_JRE_PATH_ENABLED" value="false" />
-    <option name="ALTERNATIVE_JRE_PATH" />
     <option name="PACKAGE_NAME" value="com.template" />
     <option name="MAIN_CLASS_NAME" value="com.template.DriverBasedTest" />
     <option name="METHOD_NAME" value="" />
-    <option name="TEST_OBJECT" value="class" />
     <option name="VM_PARAMETERS" value="-ea -javaagent:lib/quasar.jar" />
     <option name="PARAMETERS" value="" />
     <option name="WORKING_DIRECTORY" value="file://$PROJECT_DIR$" />
-    <option name="ENV_VARIABLES" />
-    <option name="PASS_PARENT_ENVS" value="true" />
-    <option name="TEST_SEARCH_SCOPE">
-      <value defaultName="singleModule" />
-    </option>
-    <envs />
-    <patterns />
-    <method />
+    <method v="2">
+      <option name="Make" enabled="true" />
+    </method>
   </configuration>
 </component>
