+++ cordapp-contracts-states/src/main/java/com/template/contracts/TemplateContract.java
@@ -1,4 +1,4 @@
-package com.template;
+package com.template.contracts;
 
 import net.corda.core.contracts.CommandData;
 import net.corda.core.contracts.Contract;
@@ -9,7 +9,7 @@ import net.corda.core.transactions.LedgerTransaction;
 // ************
 public class TemplateContract implements Contract {
     // This is used to identify our contract when building a transaction.
-    public static final String ID = "com.template.TemplateContract";
+    public static final String ID = "com.template.contracts.TemplateContract";
 
     // A transaction is valid if the verify() function of the contract of all the transaction's input and output states
     // does not throw an exception.
