+++ cordapp-contracts-states/src/main/java/com/template/states/TemplateState.java
@@ -1,5 +1,6 @@
-package com.template;
+package com.template.states;
 
+import com.template.contracts.TemplateContract;
 import net.corda.core.contracts.BelongsToContract;
 import net.corda.core.contracts.ContractState;
 import net.corda.core.identity.AbstractParty;
