+++ cordapp/src/main/java/com/template/flows/Responder.java
@@ -1,4 +1,4 @@
-package com.template;
+package com.template.flows;
 
 import co.paralleluniverse.fibers.Suspendable;
 import net.corda.core.flows.FlowException;
